And some more
Added game objects to game map
Added collision detection for dynamic objects
Added tree population
Fixed debug message

+++ core/src/com/agonia/game/util/Utils.java
@@ -26,6 +26,6 @@ public class Utils {
     }
 
     public static Position toTilePosition(float x, float y) {
-        return new Position(x / GameMap.TILE_SIZE, y / GameMap.TILE_SIZE);
+        return new Position((int)x / GameMap.TILE_SIZE, (int)y / GameMap.TILE_SIZE);
     }
 }
